Fix minor error and make module
