+++ src/flash_messages/mod.rs
@@ -1,0 +1,1 @@
+pub mod flash_messages;
